backend/requirements.txt
## Python 3.10.8
fastapi==0.111.1
uvicorn[standard]==0.23.2
jinja2==3.1.2
pydantic==2.5.1
python-multipart>=0.0.7
pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
opencv-python>=4.8.0,<4.9
# tensorflow==2.10.0
transformers==4.34.0   
torch==2.1.0      
pyotp==2.9.0      
# pip install "fastapi>=0.111,<0.120"
# pip install python-multipart
timm==
ultralytics